Fix concurrent access issues for metrics
This exporter effectively proxies the metrics from mongod. As it stands
now it does this using the client_golang to Set() and Reset() the
metrics. The issue is if there are concurrent scrapes then these values
are being overwritten by the next caller.

This adds locking around all of the shared metrics that are using Reset().